Dynamically determine the API port from the page URI in "production"
We also set the API's port to be 8000 in development. Unfortunately, it's
not reasonable to get this from an environment variable on the command
line when running the yarn server, and default to 8000 only when that's
not set, so we need people to modify the port in ui/config/dev.env.js,
because node.js/Vue.js.