Automatically build UI when building CLI
For some reason yarn can't be found by the Rust build.rs, under Windows,
so we need to pass in the full path to yarn for now.

Finding yarn under cmd.exe is really annoying, and PowerShell treats _any_
output on STDERR as a command failure, even if STDERR is redirected to
STDOUT. To get around this, we need to do the output redirection from
within cmd.exe, so that PowerShell doesn't see it, and still monitor
STDERR before the redirection.