Create a draft GitHub release from Travis CI when a tag is pushed
We now create `.tar.gz`, and `.zip` artifacts for Linux, macOS, and
Windows that are automatically uploaded to a draft release on GitHub
whenever we push a tag to the repository. This should help streamline the
process of releasing a new version of the tracker, especially once Travis
CI gets the Windows build/test machines sorted out.