Don't run CI for windows on Travis CI
Travis CI decided that the way to fix their windows issues with secrets
was to not filter the output of the secrets at all in the logs, instead of
fixing the filtering.

https://travis-ci.community/t/current-known-issues-please-read-this-before-posting-a-new-topic/264/21
https://github.com/travis-ci/travis-build/pull/1719

Since they're now going to output what the secrets are if we're running on
Windows, we can't really run on Windows until they provide a proper fix
for the problem.