Move map logic out of GameState & into ServerConig
There isn't really a good way to manage merging the map logic setting when
updating a new game state, so we should be storing this somewhere that we
don't need to even try to do it any more.

Originally, it was stuck on the GameState to try to minimize the number of
things that would need to be passed into a rule to have it evaluate
itself, but we're already going to have to start passing in the Dungeon
information, so there isn't really much reason to also keep trying to not
have to pass in the server settings.