Move all location availability logic out of YAML & into Rust
Unfortunately, the location availability logic is complex enough that I
don't feel like inventing a grammar for handling it in the YAML config.
This collapses the YAML side down into a single setting of "which logic
blob should I use", and everything else is handled in that particular enum
variant on the Rust side of things.