Prevent `attempt to subtract with overflow` panic doing math on unsigned numbers
`saturating_sub` clamps the lower bound of the result of the subtraction
to prevent underflowing when doing math on unsigned values. Negative
values aren't really meaningful for what we're trying to do, so it's
relatively safe to consider anything that's going to give us a negative
number the same as zero.