Independently detect Bow & Silver Arrow acquisition
`0xF5F340` only stores the state for whichever is currently
selected between the Wooden & Silver arrows for their shared inventory
slot. `0xF5F38E` is used to keep track of the bow & silver upgrade state
in a bitfield, regardless of which arrow type is the currently selected
item for the slot.

```
    Bow acquired
    |Silver Arrow acquired
    vv
   |--------|
Bit 7      0
```