Remove merging of old game state with current game state
With the update to the silver arrow tracking, this was the last of the
item states that were being pulled over from the previous game state due
to shared inventory slots. Now that we're using the actual location used
to track whether an item has been acquired, instead of which is the
currently selected one, we no longer need to merge the old game state at
all to be able to "remember" that we've seen an item.

With this, it's no longer necessary to restart the application when
starting a new randomizer run.